feat: enhance watch ratings, groups UI, and privacy management
  Major Features:
  - Upgraded rating system from integers to decimals supporting .5 increments only (7, 7.5, 8 valid; 7.2 invalid)
  - Enhanced groups pages with modern UI design (gradients, hover animations, rounded corners)
  - Added privacy indicators (lock/users icons) to watched movie cards
  - Added confirmation dialog when removing movies from watchlist

  Bug Fixes:
  - Fixed watchlist priority sorting not updating immediately after marking items high priority
  - Fixed group members displaying as Unknown by flattening username/email in DTO
  - Fixed watch card alignment issues when optional data fields are missing
  - Fixed edit watch dialog not remembering selected groups and privacy settings
  - Fixed GetByMovieIdAsync missing WatchGroups Include causing empty groupIds array
  - Fixed private watches incorrectly showing All my groups checked when editing

  Backend Changes:
  - Migrated Watch.Rating from int to double in database (EF migration)
  - Added IsPrivate and GroupIds fields to WatchResponse DTO
  - Added flattened Username/Email fields to GroupMemberResponse
  - Added WatchGroups Include to GetByMovieIdAsync query
  - Updated all rating-related DTOs to use double instead of int
  - Updated WatchMapper to extract groupIds from WatchGroups navigation property

  Frontend Changes:
  - Implemented .5 increment validation with step=0.5 for rating inputs
  - Added onPriorityChanged callback to WatchlistCard for immediate re-sorting
  - Enhanced GroupList, GroupDetail, and GroupFeed with modern styling and animations
  - Added privacy badge logic to WatchCard (Lock icon for private, Users icon for shared)
  - Fixed EditWatchDialog to check isPrivate flag before processing groupIds
  - Improved WatchCard layout with fixed heights for consistent alignment across cards
  - Changed rating parsing from parseInt to parseFloat throughout forms

+++ backend/SceneStack.API/DTOs/MovieDetailResponse.cs
@@ -32,7 +32,7 @@ public class MovieUserStatus
 {
     public int? LocalMovieId { get; set; }
     public int WatchCount { get; set; }
-    public int? LatestRating { get; set; }
+    public double? LatestRating { get; set; }
     public bool OnWatchlist { get; set; }
     public int? WatchlistItemId { get; set; }
 }
